pyproject.toml
[tool.poetry]
name = "music-genre-classification"
version = "0.1.0"
description = "CMPS344 applied data science project for music genre classification."
authors = [
    "Mario Emad Saleh Fouad <1210158@example.com>",
    "Mohamed ElSayed Shaaban <1210025@example.com>",
    "Moaz Mohamed ElSayed <1210162@example.com>",
    "Seif Wael <1210243@example.com>",
]
readme = "README.md"
packages = [{ include = "music_genre", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
joblib = "^1.4.2"
matplotlib = "^3.9.0"
mlflow = "^2.14.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
plotly = "^5.22.0"
pyyaml = "^6.0.1"
scikit-learn = "^1.5.0"
seaborn = "^0.13.2"
streamlit = "^1.36.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.0"

[tool.poetry.scripts]
music-genre-data = "music_genre.data:main"
music-genre-validate = "music_genre.data:validate_main"
music-genre-train = "music_genre.train:main"
music-genre-evaluate = "music_genre.evaluate:main"
music-genre-dashboard = "music_genre.dashboard:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=music_genre --cov-report=term-missing"
